requirements.txt
python-telegram-bot>=20.0
pyyaml>=6.0
sqlalchemy>=2.0
aiosqlite>=0.19.0
cryptography>=3.4.7
psutil>=5.8.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0 